Mitigate Timing Attack (#9)
Originally, the code did not execute all of the verifications before returning failure. This is an opportunity for a timing attack. An attacker can find which checks they are passing by measuring how long the function takes to run. This code still has some timing leaks, but they are greatly reduced.

Commits:
* Mitigate timing attacks
* Merge branch 'main' into fix/mitigate-timing