Implement Logging (#12)
For adminstrative purposes, it is often useful to log accesses to any `sudo`-like binary. This Pull Request adds this logging functionality to `sus`.

It makes a few large changes. First, it introduces `Display` implementations for the structures `Executable`, `Permission`, and `VerifyError`. This is so that they can be formatted in a human-readable way for the log file. Second, it introduces the `log` module, which provides methods to log out `to_file` or `to_write`. Accordingly, the `Request` structure is augmented with a `logger` field, and the `main` function populates it. Finally, it adds the `log` and `log_fail_msg` features so this functionality can be disabled if desired.

The `log` feature is fairly straightforward, controlling whether the enitre `log` module is used. The `log_fail_msg` is slightly more interesting. It might be unwise to log exactly what failure caused a user to not be able to authenticate. As such, this feature is disabled by default, preventing the failure message from being logged. The feature can be enabled to log the failure message too.

Many `config` options were added. Most important are `LOG_FILE_PATH` and `LOG_FILE_PERMS`. The former is the path to the log file, while the latter is the permissions the log should have. The log file will have its permissions brought into line with `LOG_FILE_PERMS` every time a log happens.

Aside from logging, some `Makefile` errors were corrected.

Closes #5

Commits:
* Added display implementation for Executable
* Added display implementation for Permission
* Add logger to Cargo.toml
* Added module boilerplate
* Added most of the scaffolding needed
* Removed logging crate
* Changed name of request
* Added basic file write
* Changed signature for Logger
* Added logging format macros
* Fine tuned messages
* Added display for VerifyError
* Changed formatting slightly
* Another minor formatting change
* Changed file name
* Added log file permissions
* Added comment to Makefile
* Added task to drop features when building
* Made errors on VerifyError optional
* Added logging feature
* Change feature name
* Added optional write reason for failure
* Modified cargo check in Makefile
* Change log file
* Fix comment