Add Userspace Binary (#16)
Ideally, the user would not have to invoke the sus kernel directly. Instead, they would do so through a userspace program that provides a nice interface. This pull request provides a basic userspae binary that does that. The interface is the same as `sudo`, but with unsupported features stripped out.

Closes #13 

Commits:
* Added startingpoint
* Did command line parsing
* Added parsing for command line options
* Added sus user configuration file
* Remove root from the groups
* Added method to get kernel arguments
* Added logic to only execute shell
* Made mesh with kernel
* Add execution
* Remove version
* Remove background
* Added documentation for main
* Pretty printed error values
* Made userspace binary optional
* Remember to capture the new artifact